Multer / code cleanup
